+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Change Log
 ### Changed
 - OpenApi generator updated to release [v4.2.3](https://github.com/OpenAPITools/openapi-generator/releases/tag/v4.2.3)
 - Enable multiple configuration files in one swagger task.
+- Tasks now contain additionalProperties setting.
+- Task configs can have all values null.
+- Generator is safe with null parameters.
 
 ## 2.0.0-alpha02 (2019-01-21)